+++ cli/docs/changelog.md
@@ -2,6 +2,7 @@
 ### Features
 * `enclave` commands also show enclave state
 * Standardized table-printing logic into a `TablePrinter` object 
+* `enclave inspect` also prints the enclave ID & state
 
 ### Fixes
 * `module exec` will attempt to update the module & API container images before running